requirements.txt
flask==2.3.3
pandas==2.0.3
boto3==1.34.60
python-dotenv==1.0.0